requirements.txt
# Azure Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Data Processing
pandas==2.2.0
pyspark==3.5.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# API Requests
requests==2.31.0